requirements.txt
python-telegram-bot==21.9
playwright==1.58.0
httpx==0.28.1
beautifulsoup4==4.13.3
python-dotenv==1.0.1
